Refactored all page objects to be non language specific - all language is now defined in i18n files in the test suite
