+++ spec/features/highlighted_response_spec.rb
@@ -3,7 +3,7 @@ require 'rails_helper'
 RSpec.describe 'Highlighted Response', type: :feature, js: true do
   include ActiveSupport::NumberHelper
   include Calculator::Test::Pages
-  let(:any_calculator_page) { Calculator::Test::En::BasePage.new }
+  let(:any_calculator_page) { Calculator::Test::BasePage.new }
 
   # Scenario: Under 61 years old view highlighted positive response
   #
