+++ spec/features/start_calculator_session_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
 RSpec.describe 'Start calculator session', type: :feature do
-  let(:start_page) { Calculator::Test::En::StartPage.new }
-  let(:marital_status_page) { Calculator::Test::En::MaritalStatusPage.new }
+  let(:start_page) { Calculator::Test::StartPage.new }
+  let(:marital_status_page) { Calculator::Test::MaritalStatusPage.new }
 
   it 'progresses from home page to the first question as suggested by the API' do
     start_page.load_page
