+++ test_common/helpers/pages.rb
@@ -2,58 +2,58 @@ module Calculator
   module Test
     module Pages
       def start_page
-        Calculator::Test::En::StartPage.new
+        Calculator::Test::StartPage.new
       end
 
       def marital_status_page
-        Calculator::Test::En::MaritalStatusPage.new
+        Calculator::Test::MaritalStatusPage.new
       end
 
       def court_fee_page
-        Calculator::Test::En::CourtFeePage.new
+        Calculator::Test::CourtFeePage.new
       end
 
       def date_of_birth_page
-        Calculator::Test::En::DateOfBirthPage.new
+        Calculator::Test::DateOfBirthPage.new
       end
 
       def disposable_capital_page
-        Calculator::Test::En::DisposableCapitalPage.new
+        Calculator::Test::DisposableCapitalPage.new
       end
 
-      # @return [Calculator::Test::En::NumberOfChildrenPage] The number of children page object
+      # @return [Calculator::Test::NumberOfChildrenPage] The number of children page object
       def number_of_children_page
-        Calculator::Test::En::NumberOfChildrenPage.new
+        Calculator::Test::NumberOfChildrenPage.new
       end
 
-      # @return [Calculator::Test::En::IncomeBenefitsPage] The benefits page object
+      # @return [Calculator::Test::IncomeBenefitsPage] The benefits page object
       def income_benefits_page
-        Calculator::Test::En::IncomeBenefitsPage.new
+        Calculator::Test::IncomeBenefitsPage.new
       end
 
-      # @return [Calculator::Test::En::TotalIncomePage] The total income page object
+      # @return [Calculator::Test::TotalIncomePage] The total income page object
       def total_income_page
-        Calculator::Test::En::TotalIncomePage.new
+        Calculator::Test::TotalIncomePage.new
       end
 
-      # @return [Calculator::Test::En::NotEligiblePage] The not eligible page object
+      # @return [Calculator::Test::NotEligiblePage] The not eligible page object
       def not_eligible_page
-        Calculator::Test::En::NotEligiblePage.new
+        Calculator::Test::NotEligiblePage.new
       end
 
-      # @return [Calculator::Test::En::FullRemissionPage] The full remission page object
+      # @return [Calculator::Test::FullRemissionPage] The full remission page object
       def full_remission_page
-        Calculator::Test::En::FullRemissionPage.new
+        Calculator::Test::FullRemissionPage.new
       end
 
-      # @return [Calculator::Test::En::PartialRemissionPage] The partial remission page object
+      # @return [Calculator::Test::PartialRemissionPage] The partial remission page object
       def partial_remission_page
-        Calculator::Test::En::PartialRemissionPage.new
+        Calculator::Test::PartialRemissionPage.new
       end
 
-      # @return [Calculator::Test::En::BasePage] The base page object which applies to any page
+      # @return [Calculator::Test::BasePage] The base page object which applies to any page
       def any_calculator_page
-        ::Calculator::Test::En::BasePage.new
+        ::Calculator::Test::BasePage.new
       end
     end
   end
