+++ test_common/messaging.rb
@@ -1,6 +1,7 @@
 require 'i18n'
 module Calculator
   module Test
+    # @private
     class Backend < ::I18n::Backend::Simple
       def initialize(messaging_dir:)
         super()
@@ -9,15 +10,19 @@ module Calculator
       end
     end
 
+    # A singleton class for translating i18n keys in the test suite.
+    # This is to allow the test suite to have its own i18n yml files so that
+    # all 'language' or 'messaging' that the test suite must use (e.g. button text to click on)
+    # is defined in a common file for each language (or even multiple files).
     class Messaging
-      def initialize(messaging_dir: File.absolute_path('../messaging', __FILE__))
-        self.backend = Backend.new(messaging_dir: messaging_dir)
-      end
-
-      def self.instance
-        Thread.current[:calculator_test_messaging_instance] ||= new
-      end
-
+      include Singleton
+      # Translates using the current locale
+      # @param [Symbol] key The key to use to lookup the text from the language file(s)
+      # @param [Symbol] locale - The locale to use (defaults to the result of #current_locale)
+      # @see #current_locale
+      # @param [Hash] options - Any options that the translation requires
+      # @return [String] The translated text
+      # @raise [::I18n::MissingTranslation] If the translation was not found
       def translate(key, locale: current_locale, **options)
         result = catch(:exception) do
           backend.translate(locale, key, options)
@@ -25,6 +30,11 @@ module Calculator
         result.is_a?(::I18n::MissingTranslation) ? raise(result) : result
       end
 
+      # Provides the current_locale which is the default for the #translate method
+      # The current locale is determined from the special environment variable
+      # 'TEST_LOCALE' - which defaults to :en
+      #
+      # @return [Symbol] The current locale
       def current_locale
         Thread.current[:calculator_test_current_locale] ||= ENV.fetch('TEST_LOCALE', 'en').to_sym
       end
@@ -33,11 +43,17 @@ module Calculator
 
       private
 
+      def initialize(messaging_dir: File.absolute_path('../messaging', __FILE__))
+        self.backend = Backend.new(messaging_dir: messaging_dir)
+      end
+
       attr_accessor :backend
     end
 
     module I18n
       extend ActiveSupport::Concern
+
+      private
 
       def t(*args)
         ::Calculator::Test::Messaging.instance.t(*args)
