+++ test_common/page_objects/calculator/not_eligible_page.rb
@@ -1,0 +1,7 @@
+module Calculator
+  module Test
+    class NotEligiblePage < BasePage
+      set_url '/calculation_result/not_eligible'
+    end
+  end
+end
