--- test_common/page_objects/en/calculator/full_remission_page.rb
@@ -1,20 +1,0 @@
-module Calculator
-  module Test
-    module En
-      class FullRemissionPage < BasePage
-        include ActiveSupport::NumberHelper
-        set_url '/calculation_result/full_remission'
-
-        def valid_for_final_positive_message?(user)
-          marital_status = user.marital_status.downcase
-          expected_header = messaging.translate "hwf_decision.final.#{marital_status}.positive.heading"
-          expected_detail = messaging.translate "hwf_decision.final.#{marital_status}.positive.detail",
-            fee: number_to_currency(user.fee, precision: 0, unit: '£'),
-            total_income: number_to_currency(user.monthly_gross_income, precision: 0, unit: '£')
-          feedback_message_with_header(expected_header) &&
-            feedback_message_with_detail(expected_detail)
-        end
-      end
-    end
-  end
-end
