--- test_common/page_objects/en/calculator/not_eligible_page.rb
@@ -1,9 +1,0 @@
-module Calculator
-  module Test
-    module En
-      class NotEligiblePage < BasePage
-        set_url '/calculation_result/not_eligible'
-      end
-    end
-  end
-end
